+++ src/extension.ts
@@ -7,6 +7,7 @@ import { registerActiveMotions } from './activeMotions';
 import { registerSelectionModifiers } from './selectionModifiers';
 import { registerSelectionFilters } from './selectionFilters';
 import { registerSelectionAlignments } from './selectionAlignment';
+import { registerSymmetricModifiers } from "./symmetricModifiers";
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -26,6 +27,7 @@ export function activate(context: vscode.ExtensionContext) {
     registerSelectionModifiers(context);
     registerSelectionFilters(context);
     registerSelectionAlignments(context);
+    registerSymmetricModifiers(context);
 }
 
 // this method is called when your extension is deactivated